refactor: update button variants to the lib folder  and add missing commas
